Add stats callback plugin
This plugin can be used to collect statistics about an Ansible run and
write them to a file in JSON format. The statistics currently include
the count and names of failed and unreachable hosts.

They also include whether execution stopped due to a lack of remaining
hosts. This last signal can be helpful if trying to determine whether
Ansible completed successfully for some but not all hosts, e.g. when
attempting to handle unreachable hosts. Unfortunately the
v2_playbook_on_no_hosts_remaining callback is not reliably fired, as
outlined in [1]. A fix for this has been proposed in [2].

[1] https://github.com/ansible/ansible/issues/81549
[2] https://github.com/ansible/ansible/pull/81550

Change-Id: I6eedc23726f47e88b50efbaa80db33965aa58e1a